feat: 모달 구현 및 이미지 업로드 구현 / 진행 중 중간 커밋
