+++ src/Pages/Admin/Employee/Employee.module.css
@@ -102,7 +102,35 @@
     width: auto;
   }
 
-  .head__position {
-    display: flex;
+  .head__birthday,
+  .birthday {
+    display: none;
+  }
+
+  .head__joinday,
+  .joinday {
+    display: none;
+  }
+
+  .head__contact,
+  .contact {
+    display: none;
+  }
+}
+/* 타블렛 */
+@media (min-width: 600px) and (max-width: 1023px) {
+  .head__birthday,
+  .birthday {
+    display: none;
+  }
+
+  .head__joinday,
+  .joinday {
+    display: none;
+  }
+
+  .table__content > .profileimg {
+    width: 80px;
+    height: 80px;
   }
 }
