+++ src/Pages/Admin/Employee/Modal/ModalStyle.css
@@ -1,0 +1,8 @@
+.Employee-edit__content {
+  box-shadow: 0 4px 8px rgba(0, 0, 0, 0.2), 0 6px 20px rgba(0, 0, 0, 0.19);
+}
+
+.modal button {
+  width: 100%;
+  margin-top: 16px;
+}
